frontend-minimal (#19)
<!-- Provide a general summary of your changes in the Title above -->

added the very minimal front end, including homepage, submit recipe
page, a recipe page
<!-- Describe your changes in detail -->

for the front end, i added an app.tsx file, pages folder with hompage,
create recipe page, and all recipe page. **needs to be hooked up with
backend, unsure how to do that so please request changes with how if
possible**. all are typescript files.

<!-- This project only accepts pull requests related to open issues -->
<!-- If suggesting a new feature or change, please discuss it in an
issue first -->
<!-- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!-- Please link to the issue here: -->

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->
accomplished our goal of sprint one, and being able to interact with the
app with a simple front-end and backend
<!-- If it fixes an open issue, please link to the issue here. -->

https://github.com/wandrew004/team-charles-app/issues/13#event-16811400763

## How Has This Been Tested?

<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran
to -->
<!-- see how your change affects other areas of the code, etc. -->
<!-- If no tests were run, reply with N/A -->
N/A

## Screenshots (if appropriate):

---------

Co-authored-by: angiechang <killmenow107329@gmail.com>
Co-authored-by: rmdluo <rmdluo@gmail.com>