new_recipe_frontend (#40)
<!-- Provide a general summary of your changes in the Title above -->

revamp of frontend to fit the design of figma wireframe. 

<!-- Describe your changes in detail -->

I kept the same functionality as the old form, but added new frontend
elements and styling with tailwind. I still need to update the static
quantity count but everything else should be connected to the backend.
also added a public folder to hold images and assets.

<!-- This project only accepts pull requests related to open issues -->
<!-- If suggesting a new feature or change, please discuss it in an
issue first -->
<!-- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!-- Please link to the issue here: -->
this relates to the following issue:
https://github.com/wandrew004/team-charles-app/issues/24

<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->

no tests have been run yet. 
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran
to -->
<!-- see how your change affects other areas of the code, etc. -->
<!-- If no tests were run, reply with N/A -->

## Screenshots (if appropriate):

---------

Co-authored-by: angiechang <killmenow107329@gmail.com>
Co-authored-by: rmdluo <rmdluo@gmail.com>